Added admin and made them un-deletable and un-removable from admin.

+++ estore-ui/estore-ui/src/app/delete-account/delete-account.component.ts
@@ -27,6 +27,10 @@ export class DeleteAccountComponent {
   }
 
   delete(): void {
+    if (this.user.email == 'admin') {
+      alert("this account is not allowed to be deleted");
+      return;
+    }
     this.userService.deleteUser(this.user.id)
     .subscribe(() => {
         console.log(`Deleted user with id=${this.user.id}`);
